# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0",
    "title": "Web Bootcamp Todo API Documentation"
  },
  "servers": [
    {
      "url": "https://api.todo.mrsunboy.com",
      "description": "Development server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "paths": {
    "/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Login",
        "description": "Get access token",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
    "/todo": {
      "get": {
        "tags": [
          "todo"
        ],
        "summary": "Get todo list",
        "description": "Get all todos",
        "operationId": "getTodo",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "post": {
        "tags": [
          "todo"
        ],
        "summary": "Post todo",
        "description": "Post a todo",
        "operationId": "postTodo",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "type": "object",
                "required": [
                  "text"
                ],
                "properties": {
                  "text": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "delete": {
        "tags": [
          "todo"
        ],
        "summary": "Delete todo",
        "description": "Delete a todo",
        "operationId": "deleteTodo",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "schema": {
              "type": "string",
              "format": "number"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}